day10: simplify
Parse now returns an iterator.

We got rid of all the CPU state and this solution is done in a single
iterator chain from string to result.

One iterator chain per part, so technically we parse twice.

The big difference is that we don't need to emulate a CPU. noop have no
observable effect, so they are equivalent to an "add 0". addx takes two
cyles, so they are equivalent to "add 0, add X".

X register state is done in an iterator scan() pass.

Taking this simplification into account, the only issue is to have a
proper cycle counter. The pipeline length means our simplification needs
to be offsets by two in part 1, and that we need to insert one at the
front and drop the last value in part 2.

We do lose some tests, but there's less to test now that there is no
CPU + cycle emulation.