day16: reimplement part 2 with a bitmask technique and single full iteration
This is inspired by other solutions and only needs a single full
exploration, no dual-state.
Potential valves (with flow) are indexed, and then used as part of a
bitmap so that we can find after the iteration the optimal path for both
the elephant and self.

With this different algorithm, we can now pass even the community test
cases. The ValveName implementation, the path finding with memoization
are kept. (although it's now obvious that Floyd-Warshall would be better
here).